fix sql query
